[plugin.video.gq] 3.0.2
