+++ plugin.video.gq/README.txt
@@ -3,11 +3,5 @@ plugin.video.gq
 
 Kodi Addon for GQ website
 
+Version 3.0.2 - website changes
 Version 3.0.1 - Isengard - separate scraper for future functions
-version 1.0.1 initial release
-1.0.2 - added HD resolution select setting
-1.0.3 - changed static strings to translatable strings
-1.0.4 - fixed the fix above
-1.0.5 - website changes
-1.0.6 - website changes
-1.0.7 - Added metadata and views
