--- plugin.video.gq/resources/settings.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="utf-8" standalone="yes"?>
-<settings>
-
-	<category label="30050">
-		<setting type="sep" />
-           <setting id="sub_enable" type="bool" label="30051" default="false"/>
-		<setting type="sep" />
-	</category>
-
-  <category label="30000">
-    <setting type="sep" />
-    <setting id="vid_res" type="enum" label="30000" lvalues="30001|30002" default="1" />
-    <setting type="sep" />
-  </category>
-
-  <category label="30073">
-		<setting type="sep" />
-		<setting id='enable_views' type ="bool" label="30070" default="false"/>
-		<setting id="default_view" type="text" label="30071" default="57" enable="!eq(-1,false)"/>
-		<setting id="episode_view" type="text" label="30072" default="56" enable="!eq(-2,false)"/>
-		<setting type="sep" />
-  </category>
-
-
-
-	<category label="30030">
-		<setting type="sep" />
-		<setting id="us_proxy_enable" type="bool" label="30042" default="false"/>
-		<setting id="us_proxy" type="text" label="30040" default=""/>
-		<setting id="us_proxy_port" type="text" label="30041" default=""/>
-		<setting id="us_proxy_user" type="text" label="30043" default=""/>
-		<setting id="us_proxy_pass" type="text" label="30044" default=""/>
-		<setting type="sep" />
-	</category>
-</settings>
